adrressing comments
